international transfer completed

+++ infrastructure/payment_processor/chimoney/types.go
@@ -25,7 +25,7 @@ type InternationalPaymentRequestPayload struct {
 	DestinationCountry string `json:"countryToSend"`
 	BankCode string `json:"account_bank"`
 	AccountNumber string `json:"account_number"`
-	ValueInUSD uint64 `json:"valueInUSD"`
+	ValueInUSD float32 `json:"valueInUSD"`
 }
 
 type InternationalPaymentRequestResponseDataPayload struct {
